모래시계 타이머 동작 완성
