--- src/app/hourglass-example/page.tsx
@@ -1,10 +1,0 @@
-'use client';
-
-import React from 'react';
-import Hourglass from '../../components/hourglass/Hourglass';
-
-const Index: React.FC = () => {
-  return <Hourglass />;
-};
-
-export default Index;
